to view by default, accepts options to assign to view model
